+++ packages/fixtures/src/index.ts
@@ -1,2 +1,0 @@
-export * from './fixtures/testWithSynpress'
-export * from './utils/getExtensionId'