+++ packages/test/src/index.ts
@@ -1,0 +1,1 @@
+export { default as testWithSynpress } from "./testWithSynpress";