+++ wallets/ethereum-wallet-mock/src/utils/index.ts
@@ -1,2 +1,2 @@
-export { default as mockEthereum } from './mockEthereum'
-export * from './constants'
+export { default as mockEthereum } from "./mockEthereum";
+export * from "./constants";