+++ wallets/ethereum-wallet-mock/test/synpress.ts
@@ -1,0 +1,4 @@
+import { testWithSynpress } from "@synthetixio/synpress-test";
+import { ethereumWalletMockFixtures } from "../src";
+
+export default testWithSynpress(ethereumWalletMockFixtures);