+++ wallets/metamask/test/e2e/openSettings.spec.ts
@@ -1,0 +1,19 @@
+import { testWithSynpress } from '@synthetixio/synpress-test'
+import { MetaMask, metaMaskFixtures } from '../../src'
+
+import basicSetup from '../wallet-setup/basic.setup'
+
+const test = testWithSynpress(metaMaskFixtures(basicSetup))
+
+const { expect } = test
+
+test('should open settings', async ({ context, metamaskPage }) => {
+  const metamask = new MetaMask(context, metamaskPage, basicSetup.walletPassword)
+
+  await metamask.openSettings()
+
+  const SidebarMenus = metamask.homePage.selectors.settings.SettingsSidebarMenus
+  await expect(
+    metamaskPage.locator(metamask.homePage.selectors.settings.sidebarMenu(SidebarMenus.General))
+  ).toBeVisible()
+})