+++ release/CHANGELOG.md
@@ -7,7 +7,7 @@
 - Alpha Release
 - Updated dependencies
   - @synthetixio/synpress-core@0.0.1-alpha.3
-  - @synthetixio/synpress-fixtures@0.0.1-alpha.3
+  - @synthetixio/synpress-test@0.0.1-alpha.3
   - @synthetixio/synpress-metamask@0.0.1-alpha.3
 
 ## 4.0.0-alpha.2
@@ -17,7 +17,7 @@
 - Alpha Release
 - Updated dependencies
   - @synthetixio/synpress-core@0.0.1-alpha.2
-  - @synthetixio/synpress-fixtures@0.0.1-alpha.2
+  - @synthetixio/synpress-test@0.0.1-alpha.2
   - @synthetixio/synpress-metamask@0.0.1-alpha.2
 
 ## 4.0.0-alpha.1
@@ -27,7 +27,7 @@
 - Alpha Release
 - Updated dependencies
   - @synthetixio/synpress-core@0.0.1-alpha.1
-  - @synthetixio/synpress-fixtures@0.0.1-alpha.1
+  - @synthetixio/synpress-test@0.0.1-alpha.1
   - @synthetixio/synpress-metamask@0.0.1-alpha.1
 
 ## 4.0.0-alpha.0